backend/requirements.txt
fastapi
google-genai
requests
pillow
uvicorn